added PaymentsService and renamed project ProductServcice -> ProductsService, added core components also
