+++ OrdersService/src/main/java/com/appsdeveloperblog/estore/OrdersService/query/OrderEventsHandler.java
@@ -32,7 +32,6 @@ public class OrderEventsHandler {
 		}catch (IllegalArgumentException e) {
 			e.printStackTrace();
 		}
-		
 	}
 	
 }
