+++ PaymentsService/src/main/java/com/appsdeveloperblog/estore/PaymentsService/core/data/PaymentsRepository.java
@@ -1,0 +1,7 @@
+package com.appsdeveloperblog.estore.PaymentsService.core.data;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PaymentsRepository extends JpaRepository<PaymentEntity, String> {
+
+}
