+++ PaymentsService/src/main/resources/application.properties
@@ -1,0 +1,21 @@
+# random port number
+server.port=0
+
+# name of the application
+spring.application.name=payments-service
+
+
+# h2 jpa configuration
+spring.datasource.url=jdbc:h2:file:~/payments;AUTO_SERVER=true
+spring.datasource.username=root
+spring.datasource.password=sandeep
+spring.datasource.driver-class-name=org.h2.Driver
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.jpa.hibernate.ddl-auto=update
+spring.h2.console.enabled=true
+spring.h2.console.settings.web-allow-others=true
+
+server.error.include-message=always
+server.error.include-binding-errors=always
+
+axon.eventhandling.processors.payment-group.mode=subscribing
