+++ PaymentsService/src/test/java/com/appsdeveloperblog/estore/PaymentsService/PaymentsServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.appsdeveloperblog.estore.PaymentsService;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PaymentsServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
