--- ProductService/src/main/java/com/appsdeveloperblog/estore/ProductService/core/data/ProductLookupRepository.java
@@ -1,8 +1,0 @@
-package com.appsdeveloperblog.estore.ProductService.core.data;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductLookupRepository extends JpaRepository<ProductLookupEntity, String> {
-	
-	ProductLookupEntity findByProductIdOrTitle(String productId, String title);
-}
