--- ProductService/src/main/java/com/appsdeveloperblog/estore/ProductService/core/data/ProductsRepository.java
@@ -1,10 +1,0 @@
-package com.appsdeveloperblog.estore.ProductService.core.data;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductsRepository extends JpaRepository<ProductEntity, String> {
-	
-	ProductEntity findByProductId(String productId);
-	ProductEntity findByProductIdOrTitle(String productId, String title);
-
-}
