--- ProductService/src/main/java/com/appsdeveloperblog/estore/ProductService/core/errorhandling/ProductsServiceEventsErrorHandler.java
@@ -1,16 +1,0 @@
-package com.appsdeveloperblog.estore.ProductService.core.errorhandling;
-
-import org.axonframework.eventhandling.EventMessage;
-import org.axonframework.eventhandling.EventMessageHandler;
-import org.axonframework.eventhandling.ListenerInvocationErrorHandler;
-
-public class ProductsServiceEventsErrorHandler implements ListenerInvocationErrorHandler {
-
-	@Override
-	public void onError(Exception exception, EventMessage<?> event, EventMessageHandler eventHandler) throws Exception {
-		
-		throw exception;
-		
-	}
-
-}
