--- ProductService/src/main/java/com/appsdeveloperblog/estore/ProductService/query/FindProductsQuery.java
@@ -1,5 +1,0 @@
-package com.appsdeveloperblog.estore.ProductService.query;
-
-public class FindProductsQuery {
-
-}
