--- ProductService/src/main/java/com/appsdeveloperblog/estore/ProductService/query/rest/ProductRestModel.java
@@ -1,15 +1,0 @@
-package com.appsdeveloperblog.estore.ProductService.query.rest;
-
-import java.math.BigDecimal;
-
-import lombok.Data;
-
-@Data
-public class ProductRestModel {
-	
-	private String productId;
-	private String title;
-	private BigDecimal price;
-	private Integer quantity;
-	
-}
