+++ ProductService/src/main/java/com/appsdeveloperblog/estore/ProductsService/core/data/ProductLookupRepository.java
@@ -1,0 +1,8 @@
+package com.appsdeveloperblog.estore.ProductsService.core.data;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductLookupRepository extends JpaRepository<ProductLookupEntity, String> {
+	
+	ProductLookupEntity findByProductIdOrTitle(String productId, String title);
+}
