+++ ProductService/src/main/java/com/appsdeveloperblog/estore/ProductsService/core/errorhandling/ProductsServiceEventsErrorHandler.java
@@ -1,0 +1,16 @@
+package com.appsdeveloperblog.estore.ProductsService.core.errorhandling;
+
+import org.axonframework.eventhandling.EventMessage;
+import org.axonframework.eventhandling.EventMessageHandler;
+import org.axonframework.eventhandling.ListenerInvocationErrorHandler;
+
+public class ProductsServiceEventsErrorHandler implements ListenerInvocationErrorHandler {
+
+	@Override
+	public void onError(Exception exception, EventMessage<?> event, EventMessageHandler eventHandler) throws Exception {
+		
+		throw exception;
+		
+	}
+
+}
