+++ ProductService/src/main/java/com/appsdeveloperblog/estore/ProductsService/core/events/ProductCreatedEvent.java
@@ -1,0 +1,19 @@
+package com.appsdeveloperblog.estore.ProductsService.core.events;
+
+import java.math.BigDecimal;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ProductCreatedEvent {
+
+	private String productId;
+	private String title;
+	private BigDecimal price;
+	private Integer quantity;
+	
+}
