+++ ProductService/src/main/java/com/appsdeveloperblog/estore/ProductsService/query/FindProductsQuery.java
@@ -1,0 +1,5 @@
+package com.appsdeveloperblog.estore.ProductsService.query;
+
+public class FindProductsQuery {
+
+}
